fix: Don't put a request in a component that can be rendered hundred times
